Working On Lingering Fire Issue
There's still a problem with the fire system where one to two tiles will continue burning even when in the finished state. I'm very close to solving this issue but am struggling to understand what is wrong in my code. My conclusion is that Unity is evil and also hates me, which is fine because the feeling is mutual at the moment